requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
autopep8==1.7.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
decorator==5.1.1
Django==4.1.2
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
executing==1.1.1
gunicorn==20.1.0
idna==3.4
ipython==8.5.0
jedi==0.18.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
postmarker==1.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
Pygments==2.13.0
PyJWT==2.5.0
python-http-client==3.3.7
pytz==2022.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
starkbank-ecdsa==2.1.0
toml==0.10.2
traitlets==5.4.0
urllib3==1.26.12
wcwidth==0.2.5
